Chore: Rename lefts/rights to left taps / right taps
