+++ ios/Approach/Sources/StatisticsLibrary/Trackable/FirstRoll/Statistics+LeftTaps.swift
@@ -2,24 +2,24 @@ import ModelsLibrary
 import StringsLibrary
 
 extension Statistics {
-	public struct Rights: Statistic, TrackablePerFirstRoll, CountingStatistic {
-		public static var title: String { Strings.Statistics.Title.rights }
+	public struct LeftTaps: Statistic, TrackablePerFirstRoll, CountingStatistic {
+		public static var title: String { Strings.Statistics.Title.leftTaps }
 		public static var category: StatisticCategory { .taps }
 		public static var isEligibleForNewLabel: Bool { false }
 		public static var preferredTrendDirection: StatisticTrendDirection? { .downwards }
 
-		private var rights = 0
+		private var lefts = 0
 		public var count: Int {
-			get { rights }
-			set { rights = newValue }
+			get { lefts }
+			set { lefts = newValue }
 		}
 
 		public init() {}
-		init(rights: Int) { self.rights = rights }
+		init(lefts: Int) { self.lefts = lefts }
 
 		public mutating func adjust(byFirstRoll roll: Frame.OrderedRoll, configuration: TrackablePerFrameConfiguration) {
-			if roll.roll.pinsDowned.isRight {
-				rights += 1
+			if roll.roll.pinsDowned.isLeft {
+				lefts += 1
 			}
 		}
