--- ios/Approach/Tests/StatisticsLibraryTests/Trackable/FirstRoll/Statistics+RightsSparedTests.swift
@@ -1,165 +1,0 @@
-import Dependencies
-import ModelsLibrary
-@testable import StatisticsLibrary
-import XCTest
-
-final class RightsSparedTests: XCTestCase {
-	func testAdjust_ByFramesWithRightsSpared_Adjusts() {
-		let statistic = create(
-			statistic: Statistics.RightsSpared.self,
-			adjustedByFrames: [
-				Frame.TrackableEntry(
-					index: 0,
-					rolls: [
-						.init(index: 0, roll: .init(pinsDowned: [.rightThreePin, .headPin, .leftThreePin, .leftTwoPin])),
-						.init(index: 1, roll: .init(pinsDowned: [.rightTwoPin])),
-					]
-				),
-				Frame.TrackableEntry(
-					index: 1,
-					rolls: [
-						.init(index: 0, roll: .init(pinsDowned: [.rightThreePin, .headPin, .leftThreePin, .leftTwoPin])),
-						.init(index: 1, roll: .init(pinsDowned: [])),
-						.init(index: 2, roll: .init(pinsDowned: [.rightTwoPin])),
-					]
-				),
-			],
-			withFrameConfiguration: .default
-		)
-
-		AssertPercentage(statistic, hasNumerator: 1, withDenominator: 2, formattedAs: "50% (1)")
-	}
-
-	func testAdjust_ByFramesWithoutLeftsSpared_DoesNotAdjust() {
-		let statistic = create(
-			statistic: Statistics.RightsSpared.self,
-			adjustedByFrames: [
-				Frame.TrackableEntry(
-					index: 0,
-					rolls: [
-						.init(index: 0, roll: .init(pinsDowned: [.rightThreePin, .headPin, .leftThreePin, .leftTwoPin])),
-						.init(index: 1, roll: .init(pinsDowned: [])),
-					]
-				),
-				Frame.TrackableEntry(
-					index: 1,
-					rolls: [
-						.init(index: 0, roll: .init(pinsDowned: [.rightThreePin, .headPin, .leftThreePin, .leftTwoPin])),
-						.init(index: 1, roll: .init(pinsDowned: [])),
-						.init(index: 2, roll: .init(pinsDowned: [.rightTwoPin])),
-					]
-				),
-			],
-			withFrameConfiguration: .default
-		)
-
-		AssertPercentage(statistic, hasNumerator: 0, withDenominator: 2, formattedAs: "0%", overridingIsEmptyExpectation: true)
-	}
-
-	func testAdjust_ByFramesWithRightSplitsWithBonusSpared_WithBonusEnabled_Adjusts() {
-		let statistic = create(
-			statistic: Statistics.RightSplitsSpared.self,
-			adjustedByFrames: [
-				Frame.TrackableEntry(
-					index: 0,
-					rolls: [
-						.init(index: 0, roll: .init(pinsDowned: [.headPin, .rightThreePin, .leftTwoPin])),
-						.init(index: 1, roll: .init(pinsDowned: [.leftThreePin, .rightTwoPin])),
-					]
-				),
-				Frame.TrackableEntry(
-					index: 1,
-					rolls: [
-						.init(index: 0, roll: .init(pinsDowned: [.headPin, .rightThreePin, .leftTwoPin])),
-						.init(index: 1, roll: .init(pinsDowned: [])),
-					]
-				),
-				Frame.TrackableEntry(
-					index: 2,
-					rolls: [
-						.init(index: 0, roll: .init(pinsDowned: [.headPin, .leftThreePin, .rightTwoPin])),
-						.init(index: 1, roll: .init(pinsDowned: [.leftTwoPin, .rightThreePin])),
-					]
-				),
-				Frame.TrackableEntry(
-					index: 3,
-					rolls: [
-						.init(index: 0, roll: .init(pinsDowned: [.headPin, .leftThreePin, .rightTwoPin])),
-						.init(index: 1, roll: .init(pinsDowned: [])),
-					]
-				),
-				Frame.TrackableEntry(
-					index: 4,
-					rolls: [
-						.init(index: 0, roll: .init(pinsDowned: [.headPin])),
-					]
-				),
-				Frame.TrackableEntry(
-					index: 5,
-					rolls: [
-						.init(index: 0, roll: .init(pinsDowned: [.headPin, .rightThreePin, .leftTwoPin])),
-						.init(index: 1, roll: .init(pinsDowned: [])),
-					]
-				),
-			],
-			withFrameConfiguration: .init(countHeadPin2AsHeadPin: false, countSplitWithBonusAsSplit: true)
-		)
-
-		AssertPercentage(statistic, hasNumerator: 1, withDenominator: 3, formattedAs: "33.3% (1)")
-	}
-
-	func testAdjust_ByFramesWithRightSplitsWithBonusSpared_WithBonusDisabled_DoesNotAdjust() {
-		let statistic = create(
-			statistic: Statistics.RightSplitsSpared.self,
-			adjustedByFrames: [
-				Frame.TrackableEntry(
-					index: 0,
-					rolls: [
-						.init(index: 0, roll: .init(pinsDowned: [.headPin, .rightThreePin, .leftTwoPin])),
-						.init(index: 1, roll: .init(pinsDowned: [.leftThreePin, .rightTwoPin])),
-					]
-				),
-				Frame.TrackableEntry(
-					index: 1,
-					rolls: [
-						.init(index: 0, roll: .init(pinsDowned: [.headPin, .rightThreePin, .leftTwoPin])),
-						.init(index: 1, roll: .init(pinsDowned: [])),
-					]
-				),
-				Frame.TrackableEntry(
-					index: 2,
-					rolls: [
-						.init(index: 0, roll: .init(pinsDowned: [.headPin, .leftThreePin, .rightTwoPin])),
-						.init(index: 1, roll: .init(pinsDowned: [.leftTwoPin, .rightThreePin])),
-					]
-				),
-				Frame.TrackableEntry(
-					index: 3,
-					rolls: [
-						.init(index: 0, roll: .init(pinsDowned: [.headPin, .leftThreePin, .rightTwoPin])),
-						.init(index: 1, roll: .init(pinsDowned: [])),
-					]
-				),
-				Frame.TrackableEntry(
-					index: 4,
-					rolls: [
-						.init(index: 0, roll: .init(pinsDowned: [.headPin])),
-					]
-				),
-			],
-			withFrameConfiguration: .init(countHeadPin2AsHeadPin: false, countSplitWithBonusAsSplit: false)
-		)
-
-		AssertPercentage(statistic, hasNumerator: 0, withDenominator: 0, formattedAs: "0%")
-	}
-
-	func testAdjustBySeries_DoesNothing() {
-		let statistic = create(statistic: Statistics.RightsSpared.self, adjustedBySeries: Series.TrackableEntry.mocks)
-		AssertPercentage(statistic, hasNumerator: 0, withDenominator: 0, formattedAs: "0%")
-	}
-
-	func testAdjustByGame_DoesNothing() {
-		let statistic = create(statistic: Statistics.RightsSpared.self, adjustedByGames: Game.TrackableEntry.mocks)
-		AssertPercentage(statistic, hasNumerator: 0, withDenominator: 0, formattedAs: "0%")
-	}
-}
